+++ class/map/CaseMap.hpp
@@ -6,7 +6,7 @@
 /*   By: lflandri <lflandri@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/21 12:06:41 by lflandri          #+#    #+#             */
-/*   Updated: 2023/02/23 18:58:52 by lflandri         ###   ########.fr       */
+/*   Updated: 2023/03/01 16:49:14 by lflandri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ private:
 
 	unsigned int	x;
 	unsigned int	y;
-	unsigned int	z;
+	float	z;
 	std::string		type;
 	sf::Sprite		*img = NULL;
 
@@ -42,8 +42,8 @@ public:
 	unsigned int getY() const;
 	void setX(unsigned int x);
 	void setY(unsigned int y);
-	unsigned int getZ() const;
-	void setZ(unsigned int z);
+	float getZ() const;
+	void setZ(float z);
 	const std::string & getType() const;
 	void setType(std::string type);
 	sf::Sprite * getImg() const;
